LayoutTests/crypto/subtle/rsa-import-pkcs8-invalid-length.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test importing a PKCS8 RSA-OAEP key with invalid length byte causing buffer overread");

// This test verifies that importing a malformed PKCS8 key where the final
// length byte has a large value (0xFF) is rejected gracefully.
// Without proper bounds checking, this would cause subspan() to attempt
// accessing memory far beyond the buffer.
//
// Key structure (22 bytes):
// - Offset 0: 0x30 (SequenceMark)
// - Offset 1: 0x00 (length byte, bytesUsedToEncodedLength returns 1)
// - Offset 2-20: 19 placeholder bytes (Version + RSAOIDHeader + OctetStringMark)
// - Offset 21: 0x83 (length byte indicating 3 more bytes encode the length)
//
// headerSize computation:
// 1. headerSize = 1
// 2. Check: 22 >= 2 (pass)
// 3. headerSize = 1 + 1 + 3 + 15 + 1 = 21
// 4. Check: 22 >= 22 (pass)
// 5. bytesUsedToEncodedLength(0x83) = 131 - 128 + 1 = 4
//    headerSize = 21 + 4 = 25
// 6. subspan(25) on 22-byte buffer -> buffer overread without bounds check

var extractable = true;

// 22 bytes: 30 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 83
var invalidLengthPkcs8Key = hexStringToUint8Array("30000000000000000000000000000000000000000083");

// Should reject with DataError, not crash or read beyond buffer
shouldReject('crypto.subtle.importKey("pkcs8", invalidLengthPkcs8Key, {name: "RSA-OAEP", hash: "sha-256"}, extractable, ["decrypt"])');

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
